fix border offsets separately from disallowed areas
The disallowed areas are given in the coordinate system of the left extruder (?),
whereas the global border is defined as the union of the extruder borders of each extruder.
In order to retrieve the border of a single extruder, we mut perform some intersection to get it out of the global border.